feat: add confirmation on remove operation

+++ src/commands/remove.ts
@@ -4,6 +4,8 @@ import { Database as db } from '../utils';
 export async function remove(cmd: any) {
 	let { id, all } = cmd;
 
+	let sure = false;
+
 	if (all) {
 		const list = db.getAll();
 
@@ -11,9 +13,24 @@ export async function remove(cmd: any) {
 			return console.warn('\nThere is nothing to remove!\n');
 		}
 
+		console.table(list);
+
+		sure =
+			(await input({
+				message: 'Are you sure these above the correct information? (y/n)',
+				validate: (value) => value === 'y' || value === 'n',
+			})) === 'y'
+				? true
+				: false;
+
+		if (!sure) {
+			console.info('\nOk, exited remove operation!\n');
+			return;
+		}
+
 		db.removeAll();
 
-		console.info('\nEverything have been remove\n');
+		console.info('\nEverything have been remove!\n');
 		return;
 	}
